final renames and restoring
