+++ prismarine-viewer/README.MD
@@ -1,0 +1,5 @@
+# Prismarine Viewer
+
+Renamed to `renderer`.
+
+For more info see [CONTRIBUTING.md](../CONTRIBUTING.md).
