+++ renderer/playground/scenes/frequentUpdates.ts
