+++ renderer/playground/scenes/lightingStarfield.ts
