+++ renderer/playground/scenes/rotationIssue.ts
