+++ renderer/playground/scenes/slabsOptimization.ts
