+++ renderer/playground/scenes/transparencyIssue.ts
